feat: refactored taskEmailFormatPage to be a modal
